imx8x: fuse: skip prog confirmation for 'hwid' command
An additional (second) confirmation dialog opens when you try to
program ECC-protected words (because you can only write them once).
We want to skip this confirmation when the fuse prog is called
from the 'hwid' command, which already asks for the first confirmation
unless you skip it with '-y' parameter.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>