ccimx8x: implement new HWID fields
To avoid the dependency of a database coded in the source, we will
program some additional bits on the HWID that allows to determine
the capabilities of the variant:
 - RAM size (a value of zero serves as backwards-compatibility indicator)
 - Wi-Fi
 - Bluetooth
 - Crypto-authentication chip
 - MCA

A new OTP word is used for some of the new fields.
U-Boot creates new variable 'module_ram' that allows OS scripts
determine the new HWID fields are present and calculate the appropriate
device tree and U-Boot files.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-895