cmd_update_nand: check file magic number to select writing method
Select writing method depending on the file-to-write magic number.
When it is a UBIFS file, a UBI volume should be created in the NAND
partition and a UBI write should be performed.
When it is not, the partition should be erased and a RAW write should
be performed.

This allows writing UBIFS files that are stored in RAM and hence
don't match the *.ubifs extension condition.

https://jira.digi.com/browse/DUB-864

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>
(cherry picked from commit 19b67ca35e8292f935dd1103acda478d7102e3d5)