update: fix override of filesize variable
On the update process, the calculated size of the file is passed
to the write function.
This file size is obtained from $filesize env variable, except when
updating from RAM (where the value must be set by the user).
The OTF mechanism however was wrongly overriding this variable to
the partition size, when it shouldn't.

Fix it by using a local variable for the OTF mechanism.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-899
(cherry picked from commit c0a277eb95179894af1f7db085bf9a12287ef7e7)