mx6qp: Place enable_ipu_clock() into a common location
Commit 19011b1d98 ("MLK-22219 mx6: Place pre_misc_setting() into a
common location") moves the initialization of the QoS settings to a
common location. However, it introduces a bug that was fixed in
commit b9dc0acc7a ("MLK-22245 mx6qp: Fix issue caused by moving
pre_misc_setting()"), so this commit is a mixture of these two commits.

Enable and configure the IPU clocks and the pre-fetch engine with
the QoS settings in U-Boot so Linux with can start these engines
successfully initialized. This avoids potential IPU display under-runs
in graphics-intensive applications on i.MX6QP/DP processors.

https://jira.digi.com/browse/CCIMX6-925

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>
(cherry picked from commit 6ea7608febb1cbe5b2891d8a71b23ec334246b2e)
Signed-off-by: Hector Palacios <hector.palacios@digi.com>