imx8: sort out QXP and DX variants by checking available cores
The QXP and DX variants of the i.MX8X family identify themselves
as the same SOC type. The only way to tell them apart is to
verify if two of the cores have been disabled by fuses.
Use the SCFW API sc_rm_get_resource_owner() to determine the
number of cores.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-900