ccimx8x: update the default value of variable 'fdt_file'
The device tree files in Linux now distinguish between QXP and DX
variants, so the default value of this variable does no longer
match any file produced by the builds.

Since U-Boot does not make any distinction, we can't use the
BOARD_DEY_NAME constant either, so this commit hard codes the
default fdt_file to the QXP variant with wireless, which is the
default SOM variant shipping on the development kit.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>