trustfence: on error, always print some message
For specific errors there are no messages shown. Even though the return
code is always correct, a user handling the U-Boot console is unlikely to
check this, so change this so that if any error happens, at least a generic
error message is shown to the user.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>
(cherry picked from commit 3197f2865cf32754adfe9cf7a1d8dd6988ef5475)