bootstream: fix out-of-boundaries write
Due to commit 04b7b71f51a49, the function write_firmware 'rewinds' the data
buffer pointer and then writes to it. This is unexpected (as it changes
input data) and can cause undefined behaviour (as there is no guarantee for
that memory address being writtable).

This went unnoticed because it does not cause any problems when using
$loadaddr as data buffer (which is the most common use case). However, in
other scenarios it can cause U-Boot to crash.

To fix this, allocate a different buffer to account for the pre-padding.
This way the original input data buffer (and the data it points to) is left
unmodified.

As the data is written in mtd->writesize chunks and we have scenarios in which
pre_padding is not a multiple of mtd->writesize, we need to account for the
case in which we need to complete the pre_padding buffer with actual
bootstream data.

https://jira.digi.com/browse/DUB-909

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>
(cherry picked from commit 13148cd98f8bc4f8a16aeeb76195b49e55b705c8)
(cherry picked from commit aec6a1e8b718ad183e1ccecd0d98bb4359435c1b)