ccimx8mn_dvk: add Ethernet support
This board uses the same PHY as the CC8X SBC Pro so let's use the same
functions for power-up and reset.
The original NXP code used direct GPIO access but we are using DT and the
GPIO DM driver instead, as in the CC8X.

The only difference is a configuration on GPR1 register with a comment
that says "Use 125M anatop REF_CLK1 for ENET1, not from external".

Signed-off-by: Hector Palacios <hector.palacios@digi.com>