update: create a __weak function for calculating U-Boot offset
The offset where to program U-Boot is different per platform, but
sometimes may be different under different conditions on the same platform.
This commit creates a __weak function for the most common case and leaves
it up to the platforms to re-define a function to handle each specific
case.

While on it, rename the B0 offset of the CC8X to be the common case.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>