digi: carrier_board: print board version and ID without string or warnings
The model (SOM + carrier board) is already printed from the 'model'
property on the device tree, so there is no point in printing it again.
Instead just print the version and ID.
Also, the "WARNING" messages confuse customers so just print "undefined"
if they are not programmed.

The new banner would be:

  Board version undefined, ID undefined

or

  Board version 3, ID 129

Signed-off-by: Hector Palacios <hector.palacios@digi.com>