trustfence: fix length computation
When validating ramdisks, the size field included in the header was
used as the size of the blob to authenticate, but this is incorrect
because it does not include the size of the header itself.

This was causing that some correctly signed images were rejected as
invalid.

Note that the blob size is padded to the next 0x1000 boundary, so
this error was only reproducible when the header size made the total
size go beyond the 0x1000 boundary (0x40 / 0x1000 = 1.6 % chance).

https://jira.digi.com/browse/DUB-814

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>