Revert "net: Only call halt on a driver that has been init'ed"
Even though this commit's change is technically correct, it causes unexpected
behaviour after a "File not found" TFTP error when using the fec_mxc driver.

The ethernet interface is supposed to be halted when not in use. Whenever a
"File not found" error occurs, a "graceful stop complete" event causes the
fec_mxc driver to halt and initialize the interface. The interface is then
halted again for good, but the check in this commit was preventing the halt
from happening, leaving the interface initialized. When initializing it again
for a new TFTP transaction, it would stay in an inconsistent state until it was
properly halted and initialized again.

Remove the check so that the interface is always halted regardless of the state
of the interface in the code.

This reverts commit 68acb51f442f08152f79772af3d41ff1e4c817cd.

https://jira.digi.com/browse/DUB-933

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>