ccimx8m: spl: fix order in which MMC clocks are set
In commit d86e787f4f83f914080fd7b6a51e87ee2c04e96c we changed the order of the
MMC interfaces, but the indexes used to set the usdhc_cfg array's clocks
remained the same. This means that when the first interface is initialized, the
second interface's clock is set and viceversa. This can cause exceptions when
board_mmc_init is called.

Instead of hardcoding the indexes, use the loop counter variable.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>