ccimx8mn-dvk: add support for the PCA9450 PMIC
The 2nd spin of the ccimx8mn uses this PMIC instead of the BD71837. Apply the
necessary changes in the device tree, configuration files and SPL PMIC
initialization code. Said code is exactly the same as the one NXP uses for
their EVK, but leaving LDO4 with its default output of 0.9V instead of setting
it to 1.2V (we use it to power VDD_PHY_0V9).

The PMIC node in the device tree isn't used in the SPL's PMIC initialization
code, so support for both PMICs can co-exist in the same U-Boot image. Use
the SOM version in the HWID to decide which PMIC to initialize.

https://jira.digi.com/browse/DEL-7060

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>