ccimx8x: update: compare SECO fw and SOC revision when updating U-Boot
Implement the validate_bootloader_image() hook to check the SECO fw on the
ccimx8x. Print an error and abort the process if the SECO fw doesn't match the
SOC revision. Since we can only account for the imx-boot images that currently
exist and have no way of recognizing future versions of the SECO fw, print a
special message with instructions on how to ensure a working update if the SECO
fw cannot be recognized.

https://jira.digi.com/browse/DUB-949

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>