fastboot: fix bootloader mmc offset calculation
This fixes a regression introduced in commit 3ab65fa90762. Before that
commit the function returned an absolute MMC offset for the bootloader
to be flashed to, but that commit changed it to a block number. So the
semantics of the returned value changed and thus the consumers of this
function are using a wrong value.

This commit changes it back to an absolute offset, while still using the
common code we use in the 'update' command.

https://jira.digi.com/browse/DUB-958

Signed-off-by: Javier Viguera <javier.viguera@digi.com>