dboot: add verbosity to device tree overlays mechanism
As the operations that 'dboot' does get more complex with the addition
of the overlays mechanism, we want to add more verbosity so that the user
sees what's going on.
This commit:
 - prints a line when loading the kernel.
 - prints a line when loading the device tree.
 - prints one line per each device tree overlay file in 'overlays' or
   a message that no overlays are being applied, if empty.

For each overlay, it looks on the device tree for a property
'overlay-description' at the root node. If it finds one it prints it
to the right of the filename.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>