trustfence: authenticate fdt overlays during load
Applying FDT overlays changes dramatically the FDT file stored in RAM, so
trying to authenticate the resulting file during boot will always fail.

To avoid that, when FDT overlay support is enabled, authenticate the base FDT
and the overlays separately immediately after loading them to RAM. If every
file is authenticated correctly before applying the overlays, the resulting
FDT can be considered as fully authenticated and the boot authentication
is skipped.

https://jira.digi.com/browse/DEL-7024

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>