image: simplify image header reading
We can directly access the header in memory without verifying the
image. At this point we want to ensure the image is authenticated
regardless of its content.

Whether the image is a correct ramdisk file or not is verified
immediately afterwards.

This allows to remove some custom modifications introduced on
commit 1af643f2c5d66fd0ca5d44f2beb54f171611d814.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>