hwid: fixup DT with HWID words as u32 values
The existing fdt_fixup_hwid() function added string properties to the DT
with the different fields of the HWID.
The encrypted environment uses the HWID raw values as key modifier so we
need to have access to these from user space. Populating the raw words of
the HWID on the DT allows for a generic and scalable way to read the HWID,
which has different size and is located on different OTP bits in every
platform.

The new u32 properties are named:
  digi,hwid_0
  digi,hwid_1
  ...

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-7185
(cherry picked from commit 20609c73ffdb2450e8eae36a406d245e08526600)
Signed-off-by: Hector Palacios <hector.palacios@digi.com>