update: select mmc partition based on mmcbootpart
For updating "uboot" partition into the eMMC we can use either the BOOT1,
BOOT2 or User Data eMMC partitions. For CC8X C0 and CC8MN modules we have
to use a different start address based on the selected partition.

eMMC partition to be used is selected with 'mmcbootpart' environment variable,
while the start address to use was based on the MMC ECSD register 179.

With the default environment and ECSD registers values, the first time a
U-Boot update was performed, it was decided to use the BOOT1 partition with
the User Data partition start address, which made the module unable to boot
from eMMC. A second 'update' process was required to correct this behavior.

Change logic to select the start address to use based on the 'mmcbootpart'
environment variable too.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>