ARM: dts: ccimx8x-sbc-pro: initialize some hardware on pre-proper
Add ccimx8x-sbc-pro-u-boot.dtsi so some hardware is initialized on
pre-relocation of U-Boot proper.

Some nodes need to be probed at this stage, like the pd_dma_* nodes.
Others may come in handy the moment SPL is supported for this platform.

Because of this, CONFIG_SYS_MALLOC_F_LEN needs to be increased to 0x8000
for the system to boot correctly.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>