net: eth-uclass: skip next eth devices loading if eth0 load failed
This caused a Segmentation fault when initializing the MDIO bus for
eth1 after eth0 failed initialization (they share the MDIO bus).

Was detected when updating a device with encrypted environment with
a U-Boot with no environment encryption support. The imported
environment is encrypted and cannot be read, so the eth0 interface fails
initialization due to an unexisting "ethaddr" environment variable.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>