board: digi: remove unnecessary phy_write call from ccimx6ul/ccimx8x board code
The ccimx6ul SBC Express, ccimx6ul SBC Pro and ccimx8x SBC Express boards all
use the Microchip LAN8720A ethernet PHY. Initially, we used NXP's
mx6ul_14x14_evk board as reference, which assembles a PHY which is similar
(Microchip KSZ8081RNA), but has a different register map. The phy_write() call
in the board_phy_config() implementation of the aforementioned Digi boards is
needed for the PHY used by NXP, but not by the PHY we use. On top of that, this
call has also been known to cause failures on boards that are based on our
reference designs, so we should remove it.

Since removing this call leaves the same board_phy_config() implementation as
the default one in drivers/net/phy/phy.c, we can remove the entire function and
fall back to the default implementation.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>