arm: stm32mp: add FIP and STM32IMAGE support
By default for trusted boot with TF-A, U-Boot (u-boot-nodtb)
is located in FIP container with their device tree with
secure monitor (provided by TF-A or OP-TEE).
The FIP file is loaded by TF-A BL2 and each components is
extracted at the final location.

This patch add CONFIG_STM32MP15x_STM32IMAGE to support of
STM32 image generation for SOC STM32MP15x
when FIP container is not used (u-boot.stm32 is loaded by TF-A
as done previously to keep the backward compatibility).

Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>
Change-Id: I75c3a4359a6e42c20360fb1774a2e371a08a6270
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/184632
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/191843