ccimx6/6+/6ul: explicitly set CSF_SIZE to 0x4000 in all defconfigs
We used to set 0x4000 (the value used in i.MX6 platforms) in the Kconfig as a
fallback but the default value changed to 0x2060, so explicitly set the value
to 0x4000 when needed.

While at it, update all ccimx6/6+/6ul defconfigs with a "make savedefconfig"
and remove the CONFIG_CSF_SIZE definition from mx6_common.h.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>