env: always define env_get_offset(), and add sibling for redundant
Function env_get_offset() was defined differently when ENV_IS_EMBEDDED
was defined or not.
A better implementation is to define it always with a __weak implementation
that returns the default value (passed as parameter).

While on it, fix a typo and add also a hook for the redundant copy.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-7337