board: ccimx6ul: implement env_get_offset() depending on NAND size
Commit aea45a46fef437a96e595689de00418166ed7205 increased the U-Boot
partition without taking into account the hardcoded value of
CONFIG_ENV_OFFSET. As a result, there were issues with the location of the
environment.

This commit implements the hook env_get_offset() to return the offset of
the environment partition basing on the partition table, which varies
depending on the NAND size.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-7337