trustfence: fix authentication support to OS artifacts
Commit f926208a50 ("trustfence: Add image authentication of the different
artifacts") added support to authenticate different OS artifacts, however in
the devices with AHAB support the authentication process extracts the artifact
content in a different destination address than the source address.
To manage this behavior and avoid jumping to the artifact content with a
hardcoded header size, this commit uses a dynamic destination address that gets
modified during the authentication process.

https://jira.digi.com/browse/DEL-7371

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>