trustfence: implement AHAB SRK revocation
Though SRK revocation is configured in the CSF during the signing process,
it is required to send the COMMIT command to the SECO to consolidate the
revocation and make it permanent in the device by writing it into the
corresponding fuse word.

Modify 'trustfence revoke' AHAB implementation so it takes no argument. All
the keys in the SRK_REVOKE_MASK will be revoked simultaneously. After running
the SECO COMMIT command successfully, the currently revoked keys will be
printed.

After this point, SECO FW will always match the SRK_REVOKE_MASK value used
in the CST with the already revoked keys and will refuse to boot the device
if they do not match. It will accept a more restrictive mask, but never a less
restrictive one.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>