ccimx8mn: Add a new usb gadget node for OTG interface
Since one dts node can only bind to one DM driver in uboot, for usbotg
node, we can not use it for both DM usb host driver and DM usb gadget
driver. So a new usb gadget node is added to each usbotg node, the
original usbotg node is bind to usb host driver as default, and the
new usb gadget node is bind to usb gadet driver as default. Based on commit
22071f5a08 ("MLK-22336-2 dts: Add a new usb gadget node on imx6/7/7ulp/8/8mm")

https://jira.digi.com/browse/DUB-985

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>