trustfence: fix SRK_REVOKE word masking
SRK_REVOKE fuseword is first masked and later shifted when we program it
but it is also first masked and later shifted when we sense it.

These operations should be opposite, so reverse the sense operation to first
shift the word and then mask it. This way the mask must always be represented
aligned at bit b0.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>