trustfence: read SRK revoke mask from storage media
When the device boots from internal storage media we can read the
Second Container header flags to know in advance the SRK revoke mask
that will be permanently fused.

Present that information to the user prior to confirm proceeding with
the revocation and also abort the process when the last SRK is attempted
to be revoked.

When reading the mask from storage fails or when the device boots from
USB this information will be ignored and the user will only be informed
of the revoked keys after the fuseword is programmed.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>