board: stm32mp1: use CONFIG_SYS_MMC_ENV_DEV when available
Check whether user has explicitly defined the mmc device to use
in mmc_get_env_dev() with CONFIG_SYS_MMC_ENV_DEV.

By default, on ST boards the mmc device is selected by boot instance
device and mmc alias in device tree.

This patch allows to override this behavior for the support of customer
boards; for example when SDMMC1 is not used and ENV in mmc0=SDMMC2.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I8cbd6f92056fa2d10bf3880746d89ee0779f8e27
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/191271
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Patrice CHOTARD <patrice.chotard@st.com>