ccimx8mm: always use 33K offset when reading/writing bootloader image
On the 8M Nano, the 32K offset is only used when the bootloader image is stored
in the user data area of the eMMC, to skip a GPT partition table if one was to
exist. This offset isn't used when the bootloader is stored in the special boot
eMMC partitions, because there's no chance that they will contain a partition
table.

However, the 8M Mini always uses the same offset of 33K regardless of the
partition the image is stored in, so reflect this in the SPL and U-Boot code.

https://jira.digi.com/browse/DUB-943
https://jira.digi.com/browse/DUB-987

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
Signed-off-by: Hector Palacios <hector.palacios@digi.com>