ccimx8m: return an error code if DEK blob isn't found
The return code of the get_dek_blob() function is checked in several places to
deduce if a DEK blob is available on the storage media, but the implementation
of this function in the 8M Mini/8M Nano always returns a success code. Make
sure to return an error code of "1" when appropriate so that those checks work
as expected.

https://jira.digi.com/browse/DUB-991

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit bc235be0ff4948ad5f8e091e4e7f291187ecc371)