csf_templates: spl: unlock CAAM MID registers
The HABv4 documentation for the i.MX8M family of SOCs states that HAB versions
older than v4.4 automatically lock the CAAM's Job Ring and DECO master ID
registers on closed devices, and that NXP's secure boot implementation requires
the registers to be writable when booting the system.

The i.MX8M Mini hardware is older than the Nano's, and contains an older HAB
version (v4.3), so we need to explicitly unlock the CAAM's MID registers in the
SPL's CSF so the secure boot can work as expected. Otherwise, the OP-TEE will
hang when trying to configure the CAAM.

Add the entry to unlock the CAAM's MID registers in the SPL CSF templates.

https://jira.digi.com/browse/DUB-992

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>