stm32mp: stm32prog: remove all the header check for UART download
This patch removes the header check for UART download;
the check of checksum is not mandatory with even parity and chuck
checksum for each 256 received bytes and it is only done for
STM32 image (FSBL = TF-A BL2), not for FIT image.

This patch solve issue of duplicated 0x100 byte written with FIP header.

Fixes: 32619a4db6c9 ("stm32mp: stm32prog: add FIP header support")
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Idc6a3beb93aa6eeee965d3d12d45fa8b834c4e02
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/204037
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>