env: add mechanism to restore environment from an old location in NAND
U-Boot needs to know (precompiled) the location of the environment.
Occasionally, however, for size and geometry reasons, we need to increase
the bootloader partition, thus moving the environment partition offset.
The drawback of moving the environment is that the new U-Boot expects the
environment in the new offset, and is not able to read back the old
environment from the previous location.

This mechanism does the following:
- If a valid environment can't be read (from either of the two copies)
  instead of printing an error and using the precompiled default
  environment, if an old offset is defined, the function returns an error.
- The error is captured and the process tries to read the environment
  from the old offsets (in order up to OLD_ENV_OFFSET_LOCATIONS).
  If read at all offsets fails, the default environment is used.
  If there is a valid environment, it's read back and saved to the new
  location, except for the partition table, which might be different and
  is the one that contains the new location of the environment partition.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-7465