ccimx6ul: add old environment offsets
This allows for a recovery mechanism to try to recover the environment
from a previous location.

The environment moved from 3MiB offset to 5MiB offset on 1GB variants
when support for 4K page NANDs was added. Although only one offset is
required now (3MiB), add also the current env offset for the 1GB variant
for completeness, and in case of future moves.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-7465