board_f: round down to 64 KiB limit instead of 4 KiB for U-Boot relocation
We have observed that, on the ccimx6ul, making customizations to the U-Boot
code that increase the binary's size sometimes results in the boot process
freezing after the U-Boot relocation. Even though the exact cause of this issue
is unclear, increasing the distance between the U-Boot code and the elements
after it fixes the issue. The downside to this is that it generates a 60 KiB
section of unusable memory.

https://jira.digi.com/browse/DUB-999

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit fab28fdd2df974961283501886ba600eb789f5c4)