update: admit (and dismiss) old 'fs' parameter
When updating from a block device (mmc|usb|sata), the old syntax
admitted a file system (fs) parameter to specify the file system
format of the source block device. This allowed to use 'fatload'
or 'ext4load' to load the file for the update.
With the merge of both commands on the general 'load' command,
this was no longer required, so it was removed from the syntax.

For backwards compatibility reasons, this commit checks if
such parameter (which can only have the value 'fat' or 'ext4')
has been passed and, if so, it just dismisses it, rather than
failing.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-1001
(cherry picked from commit 6c7ea99e3a8bfcacb638c7c876e641dd803b34d9)
Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>