sign.sh: sign DCD memory block on "usb-signed" U-Boots
Create a CSF binary where both U-Boot and DCD memory blocks are signed.
On i.MX6UL platform, clearing the DCD pointer is not required.

This way we generate a "usb-signed" U-Boot that can be used with UUU.

Also, change the variable name used to generate a "usb-signed" U-Boot
to "ENABLE_USB_SIGN".

https://jira.digi.com/browse/DEL-7510

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>