ccimx6: check the correct CPU type before applying calibration
The array that contains the calibration values for CC6 variants
contains both Quad/Dual variants and DualLite variants.
Since the CPUs are different, applying the calibration for a
DualLite on a Quad/Dual variant (or viceversa) may lead to a
system hang.
This may only happen in case of a wrongly programmed variant
code on the HWID.
The additional check introduced by this commit skips the applying
of calibrated DDR values if the CPU type doesn't match with
the one reported by the HWID.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit da90051dc711f1733ffb15ea8ac2e1f573cd9d91)