ccimx8x_sbc_pro: move Android configuration to a new file
At the moment this is cosmetic, no functional change as the configs in
the new file are exactly the same to the ones removed in the "common"
file.

However, this change allows to override configuration done for DEY, as
the inclusion of the android config file is done at the end of the
device configuration, not on the 'common' file as before.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>