ccimx6ul: add support for one partition 'system' with many UBI volumes
Traditionally, Digi used a dedicated MTD partition for the main partitions
'linux', 'recovery', 'rootfs', 'update'.
This implementation has disadvantages on NANDs because it doesn't make the
best use of wear levelling.

This commit restructures the partition table as follows:
- Respect the partitions:
	bootloader
	environment
	safe
- Create a new partition called 'system' to hold the following UBI volumes:
	linux
	recovery
	rootfs
	update

The sizes of the volumes depend on the NAND size.

This support is activated by a new configuration option
ONE_PART_MANY_UBIVOLS which is disabled by default on the CC6UL.

We support two partition tables:
- small: for 256MB NAND, with smaller 'bootloader' and 'environment'
         partitions.
- big:   for the rest of variants with bigger 'bootloader' and
         'environment' partitions.

In a similar way as we generate 'mtdparts' variable when it doesn't exit
basing on the NAND size, this commit generates a new variable
'ubivolscript' with commands to:
 - erase the 'system' partition
 - create the UBI volumes

The user needs to run this script the first time to create the layout.
On the install script, this will be done by the script.

TODO:
 - dualboot support

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-7614