dboot: add support to boot system in one partition with many UBI volumes
This commit preserves the traditional syntax of the dboot command:

    	dboot <os> <media> ...

If it doesn't find an MTD partition called <os>, looks for partition
'system' and if it finds it, searches a UBI volume named <os>.
If it finds it, reads the kernel and DTB from it.

The boot arguments are different: only one 'ubi.mtd' for the system
partition and 'root=ubi0:rootfs' for the root file system.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-7614