ccimx6ul: create a script for partitioning that depends on variables
Originally, if the 'mtdparts' variable did not exist, one was created
basing on:
- the size of the NAND
- the value of $dualboot
- the value of $ubisysvols

Since this was only run once on boot, you couldn't have it modify the
partition table after changing one of the variables.

By changing this to generate a script, you can change the value of
variables, and then manually run the script 'partition_nand_linux' to
have it recreate the partition table.
Still, if the 'mtdparts' variable doesn't exist, the script will be run
by U-Boot.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>