trustfence: fix SPL build for CC8X platform
Since there is not support to trustfence in SPL for the CC8X platform, this
commit encapsulates the code to fix a build issue.

This is fixing following error:

error: 'dek_blob_spl_final_dst' undeclared

Actually, that variable is declared above, but enclosed in a 'ifdef
CONFIG_ARCH_IMX8M' guard, so the uses of that variable should be guarded
by the same define.

https://onedigi.atlassian.net/browse/DEL-7638

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>
Signed-off-by: Javier Viguera <javier.viguera@digi.com>