common: add framebuffer in chosen sub-node
In the Linux binding
Documentation/devicetree/bindings/display/simple-framebuffer.yaml
the famebuffer must be sub-nodes of chosen node.

It was not the case today in fdt_simplefb_add_node().

This patch also adds the missing #address-cells and #size-cells property
inherit from the root node as it is required by fdt_setup_simplefb_node()
and a empty "ranges" property to indicate a 1:1 translation.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Iafba8c82c16551cf84d0f6e703af5553ef7c4624
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/218967
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/220891
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
Tested-by: Yannick FERTRE <yannick.fertre@foss.st.com>
Tested-by: Patrick DELAUNAY <patrick.delaunay@foss.st.com>